move .rspec options to workflow files
