+++ .github/workflows/gem_tlc.yml
@@ -28,7 +28,12 @@ jobs:
       run: bundle exec rsmp site --type tlc --config config/simulator/tlc.yaml 2>&1 > simulator.log &
     - name: Run tests
       shell: bash
-      run: bundle exec rspec --tag '~script' spec/site
+      run: |
+        bundle exec rspec \
+        --tag '~script' \
+        --format documentation \
+        --format Details --out log/validation.log \
+        spec/site
       env:
         SITE_CONFIG: config/gem_tlc.yaml
     - name: Show validation.log
