+++ .rspec
@@ -1,3 +1,1 @@
 --require spec_helper --force-color
---format documentation
---format Details --out log/validation.log
