+++ spec/support/validator.rb
@@ -2,30 +2,6 @@ require 'rsmp'
 require 'colorize'
 require 'rspec/expectations'
 
-
-# Class used as a stream by the RSMP::Logger.
-# When RSMP::Logger writes to it, the data
-# is passed to an RSpec reporter, which
-# will in turn distribute it to the active
-# formatters, which can each write to a separate
-# file, or to the console.
-
-class ReportStream
-  def initialize rspec_reporter
-    @reporter = rspec_reporter
-  end
-
-  def puts str
-    @reporter.publish :log, message: str
-  end
-
-  def flush
-    #@reporter.flush
-  end
-end
-
-
-
 module Validator
   include RSpec::Matchers
